requirements.txt
Django==3.1.7
djangorestframework==3.12.2
psycopg2-binary>=2.8
python-decouple>=3.4
requests>=2.25.1
django-cors-headers>=3.7.0